Restore Hue Labs effects after light states
Enabling a Hue Labs effect will also automatically save the current
state to a scene, so sometimes there's a timing problem and the bridge
ends up storing the last color and brightness from Ambilight effects.
This will then cause problems later once the effect is disabled again,
and the scene is restored.

To avoid this we can restore the light states first, before enabling
the effect sensor and triggering the scene save.